\name{QPAD}
\alias{qpad}
\alias{qpad-package}
\alias{QPAD-package}
\alias{}
\title{
QPAD: Calibrating indices of avian density from non-standardized survey data
}
\description{
The analysis of large heterogeneous data sets of avian point-count surveys 
compiled across studies is hindered by a lack of analytical approaches that 
can deal with detectability and variation in survey protocols.

We reformulated removal models of avian singing rates and distance sampling 
models of the effective detection radius (EDR) to control for the effects of 
survey protocol and temporal and environmental covariates on detection 
probabilities.

These estimating procedures as described in 
Solymos et al. (2013) are implemented in the 'detect' R package.

The estimates of singing rates and effective detection distances for 
North American boreal forest songbird species is provided as part of the 
QPAD package.

Using offsets derived from these estimates can significantly reduce the 
computational burden when fitting complex models to large data sets and 
can be used with a wide range of statistical techniques for inference and 
prediction of avian densities.
}
\usage{
}
\arguments{
  \item{}{}
  \item{\dots}{other arguments passed to the functions}
}
\details{
The analysis of large heterogeneous data sets of avian point-count surveys 
compiled across studies is hindered by a lack of analytical approaches that 
can deal with detectability and variation in survey protocols.

We reformulated removal models of avian singing rates and distance sampling 
models of the effective detection radius (EDR) to control for the effects of 
survey protocol and temporal and environmental covariates on detection 
probabilities.

These estimating procedures as described in 
Solymos et al. (2013) are implemented in the 'detect' R package.

The estimates of singing rates and effective detection distances for 
North American boreal forest songbird species is provided as part of the 
QPAD package.

Using offsets derived from these estimates can significantly reduce the 
computational burden when fitting complex models to large data sets and 
can be used with a wide range of statistical techniques for inference and 
prediction of avian densities.
}
\value{

}
\references{

}
\author{
Peter Solymos, Elly Knight
}
\examples{
\dontrun{
load_BAM_QPAD(version=2)
}
}
\keyword{ manip }
\keyword{ models }
